requirements.txt
pandas
scikit-learn
joblib
pytest
mlflow
azure-ai-ml>=1.0.0


